+++ NEWS.md
@@ -1,3 +1,10 @@
+# queryup 1.0.2
+
+* Package dependencies on `RCurl` and `jsonlite` packages have been replaced 
+by a single dependency on the `httr` package.
+* When a request fails, a message with the http status code is returned.
+* Parameters `print_url` and `print_uniprot_messages` have been deprecated.
+
 # queryup 1.0.1
 
 * The function `get_uniprot_data` now works also for long queries (with more